Generator

+++ Generator.cpp
@@ -92,7 +92,7 @@ struct testCase {
 };
 
 class Solution : private testCase {
-    const int power(int n, int expo, int MOD = 1e18 + 7) {
+    const int power(int n, int expo, int64_t MOD = 1e18 + 7) {
         int ans = 1; n %= MOD;
         while (expo) {
             if (expo & 1) ans = (ans * 1ll * n) % MOD;
